requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
ujson==5.9.0
